# docs/openapi.json
{
  "components": {
    "schemas": {
      "Envelope_ListingData_": {
        "properties": {
          "cache_expires_at": {
            "format": "date-time",
            "title": "Cache Expires At",
            "type": "string"
          },
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ListingData"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "found": {
            "default": true,
            "title": "Found",
            "type": "boolean"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "partial": {
            "default": false,
            "title": "Partial",
            "type": "boolean"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "source": {
            "default": "nyaa",
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningItem"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "source_url",
          "fetched_at",
          "cache_expires_at",
          "fingerprint",
          "data"
        ],
        "title": "Envelope[ListingData]",
        "type": "object"
      },
      "Envelope_ResolveData_": {
        "properties": {
          "cache_expires_at": {
            "format": "date-time",
            "title": "Cache Expires At",
            "type": "string"
          },
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ResolveData"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "found": {
            "default": true,
            "title": "Found",
            "type": "boolean"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "partial": {
            "default": false,
            "title": "Partial",
            "type": "boolean"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "source": {
            "default": "nyaa",
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningItem"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "source_url",
          "fetched_at",
          "cache_expires_at",
          "fingerprint",
          "data"
        ],
        "title": "Envelope[ResolveData]",
        "type": "object"
      },
...
      "Envelope_TorrentDetail_": {
        "properties": {
          "cache_expires_at": {
            "format": "date-time",
            "title": "Cache Expires At",
            "type": "string"
          },
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/TorrentDetail"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "found": {
            "default": true,
            "title": "Found",
            "type": "boolean"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "partial": {
            "default": false,
            "title": "Partial",
            "type": "boolean"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "source": {
            "default": "nyaa",
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningItem"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "source_url",
          "fetched_at",
          "cache_expires_at",
          "fingerprint",
          "data"
        ],
        "title": "Envelope[TorrentDetail]",
        "type": "object"
      },
...
      "FilterMode": {
        "enum": [
          "all",
          "no_remakes",
          "trusted"
        ],
        "title": "FilterMode",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ListingData": {
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/FilterMode"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TorrentSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "media_type": {
            "$ref": "#/components/schemas/PublicMediaType"
          },
          "order": {
            "$ref": "#/components/schemas/SortOrder"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "sort": {
            "$ref": "#/components/schemas/SortMode"
          }
        },
        "required": [
          "page",
          "limit",
          "has_more",
          "filter",
          "media_type",
          "sort",
          "order",
          "items"
        ],
        "title": "ListingData",
        "type": "object"
      },
...
      "PublicMediaType": {
        "enum": [
          "all",
          "manga",
          "light_novel",
          "novel",
          "artbook",
          "unknown"
        ],
        "title": "PublicMediaType",
        "type": "string"
      },
...
      "SortMode": {
        "enum": [
          "date",
          "seeders",
          "leechers",
          "downloads",
          "size",
          "comments"
        ],
        "title": "SortMode",
        "type": "string"
      },
      "SortOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder",
        "type": "string"
      },
...
      "WarningItem": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "WarningItem",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Unofficial self-hosted JSON API for Nyaa Literature - English-translated. The upstream category is permanently locked to c=3_1. The service returns metadata, magnet links and .torrent URLs but never downloads torrent content.",
    "title": "API_Nyaa",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/latest": {
      "get": {
        "operationId": "latest_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 75,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FilterMode",
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "media_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PublicMediaType",
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortMode",
              "default": "date"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "desc"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ListingData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest",
        "tags": [
          "Search"
        ]
      }
    },
    "/search": {
      "get": {
        "operationId": "search_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 75,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FilterMode",
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "media_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PublicMediaType",
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortMode",
              "default": "date"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "desc"
            }
          },
          {
            "in": "query",
            "name": "include_details",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Details",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ListingData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "Search"
        ]
      }
    },
    "/search/resolve": {
      "get": {
        "operationId": "resolve_search_resolve_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FilterMode",
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "media_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PublicMediaType",
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "include_details",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Details",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ResolveData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve",
        "tags": [
          "Search"
        ]
      }
    },
    "/torrents/by-hash/{info_hash}": {
      "get": {
        "operationId": "torrent_by_hash_torrents_by_hash__info_hash__get",
        "parameters": [
          {
            "in": "path",
            "name": "info_hash",
            "required": true,
            "schema": {
              "title": "Info Hash",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_description",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Description",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_raw",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Raw",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_files",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Files",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "files_offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Files Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "files_limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Files Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TorrentDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Torrent By Hash",
        "tags": [
          "Torrents"
        ]
      }
    },
    "/torrents/{torrent_id}": {
      "get": {
        "operationId": "torrent_detail_torrents__torrent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "torrent_id",
            "required": true,
            "schema": {
              "title": "Torrent Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_description",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Description",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_raw",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Raw",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_files",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Files",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "files_offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Files Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "files_limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Files Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TorrentDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Torrent Detail",
        "tags": [
          "Torrents"
        ]
      }
    }
  }
}